+++ ulibgsmmap/UMGSMMAP_USSD_String.m
@@ -47,7 +47,7 @@
 
 - (NSString *)stringValue
 {
-    NSString *s = [self.asn1_data stringFromGsm7:(self.asn1_data.length * 2)];
+    NSString *s = [self.asn1_data stringFromGsm7:(int)(self.asn1_data.length * 2)];
     return s;
 }
